Resurrected the old evaluator, deleted the abstract evaluator and the interface.
